+++ src/Minifier/MinifierInterface.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace Sensiolabs\MinifyBundle\Minifier;
 
+use Sensiolabs\MinifyBundle\Minifier\Options\OptionsInterface;
+
 /**
  * @author Simon André <smn.andre@gmail.com>
  */
@@ -20,9 +22,11 @@ interface MinifierInterface
 {
     public const TYPE_CSS = 'css';
     public const TYPE_JS = 'js';
+    public const TYPE_HTML = 'html';
 
     /**
-     * @param self::TYPE_CSS|self::TYPE_JS $type
+     * @param self::TYPE_CSS|self::TYPE_JS|self::TYPE_HTML $type
+     * @param OptionsInterface|null                        $options
      */
-    public function minify(string $input, string $type): string;
+    public function minify(string $input, string $type/* , ?OptionsInterface $options = null */): string;
 }